+++ calender.h
@@ -1,4 +1,4 @@
 #ifndef __CALENDER_H__
 #define __CALENDER_H__
-void show_calender(short _year, short _month, short _day);
+void show_calender(int _year, int _month, int _day);
 #endif
